ai/requirements.txt
flask
flask-cors==3.0.10
openai==0.27.0
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
numpy
pandas
scikit-learn
gunicorn==20.1.0 